+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `BakeScreenreaderSpans` direction (minor)
+
 ## [15.0.0] - 2021-11-05
 
 * Add unstyled tables to `BakeTableBody` (minor)